fix: change jury_seed from INTEGER to BIGINT to support epoch-ms values
jurySeed values are epoch milliseconds (e.g. 1773071803978) which exceed
the PostgreSQL INTEGER max of ~2.1 billion. This caused "value out of
range for type integer" errors for 5 submissions.

- db/schema.sql: jury_seed INTEGER -> BIGINT
- migrate_submissions_from_kv.sql: cast jurySeed as ::bigint instead of ::integer

+++ db/schema.sql
@@ -218,7 +218,7 @@ CREATE TABLE submissions (
   di_partner_id     UUID REFERENCES users(id),
 
   -- Jury metadata
-  jury_seed         INTEGER,
+  jury_seed         BIGINT,
   jury_seats        INTEGER,
   internal_jury_size INTEGER,
